requirements.txt
amqp==2.1.1
APScheduler==3.3.1
beautifulsoup4==4.5.1
billiard==3.5.0.2
celery==4.0.0
cssselect==1.0.1
Django==1.10.4
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-crispy-forms==1.6.1
django-el-pagination==3.0.1
django-registration==2.2
feedparser==5.2.1
kombu==4.0.0
lxml==4.0.0
pytz==2016.7
requests==2.11.1
six==1.11.0
tzlocal==1.4
vine==1.1.3
